Se cambia de lugar el ACTION REDUCER Y STORE a la carpeta redux, se genera la funcion product by name en action y estado nuevo de productos filtrados en el reducer
