+++ client/src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import {Provider} from 'react-redux'
-import {store} from './Store/index.js'
+import {store} from './Redux/Store/index'
 
 ReactDOM.render(
   <Provider store={store}>
